A lot of major changes.
There is now a context menu when you right click on Launchy.  This gives you
the option of actually closing the program ;) as well as a new options dialog!

The options dialog was badly needed and it serves a dual purpose.  While it
handles gui the Options.cs class also handles loading/saving of options to
the registry.

The general tab of the options directory is currently empty, but should be filled
in with internet searches soon (e.g. google, weather, dictionary).

The advanced tab lets you add your own directories and filetypes!  I've wanted
this feature in Launchy for awhile now.

Anyway, I *just* got the advanced tab stuff working and it seems to be okay but
I wouldn't be surprised if a fresh install of Launchy had registry problems, I
also wouldn't be surprised if the advanced tab gui has some problems.

Josh


Todo:
1) When the options dialog closes, save the options to registry and
   perform a rescan


git-svn-id: https://launchy.svn.sourceforge.net/svnroot/launchy@9 a198a55a-610f-0410-8134-935da841aa8c