Updated weby
For starters, the IconCache now gives up if it can't find the icon cached and requests the icon to be downloaded by the main thread in the background for the next time the indexing occurs (so as to avoid any interface lag).  This still isn't working well, for instance amazon's icon will be overwritten with google's and I have no idea why the qnetworkreply's seem to have data from previous replies carried over to new replies.  I've gotten rid of the redirection stuff, as that made things complicated, but it should probably go back in, that was nice to have.

Next, I've gotten rid of the "base" and "query" aspects to weby urls.  It's just "query" now, and I've added code to update old weby configs to the new mode.

It now also updates old configs to contain suggestion urls.

git-svn-id: https://launchy.svn.sourceforge.net/svnroot/launchy@477 a198a55a-610f-0410-8134-935da841aa8c